Change readGuitarAction
Check previous address path, and if it`s guitar selection view, the guitar model is validated and saved in session, else an attempt is made to retrieve the model from the session.
It is necessary not to define the guitar model every time.
A link to roll back the form changes will be added later.
Add result view for guitar selection, empty yet.